Move Format, SignaturDecipher and VideoInfo into Provider\Youtube namespace
