+++ tests/Unit/Provider/Youtube/FormatTest.php
@@ -1,9 +1,8 @@
 <?php
 
-namespace YoutubeDownloader\Tests\Unit;
+namespace YoutubeDownloader\Tests\Unit\Provider\Youtube;
 
-use YoutubeDownloader\Format;
-use YoutubeDownloader\VideoInfo\Format as FormatInterface;
+use YoutubeDownloader\Provider\Youtube\Format;
 use YoutubeDownloader\Tests\Fixture\TestCase;
 
 class FormatTest extends TestCase
@@ -34,7 +33,7 @@ class FormatTest extends TestCase
 			'ip' => '211.12.135.54',
 		];
 
-		$video_info = $this->createMock('\\YoutubeDownloader\\VideoInfo');
+		$video_info = $this->createMock('\\YoutubeDownloader\\Provider\\Youtube\\VideoInfo');
 		$video_info->method('getVideoId')->willReturn('ScNNfyq3d_w');
 
 		$config = [];
@@ -42,6 +41,7 @@ class FormatTest extends TestCase
 		$format = Format::createFromArray($video_info, $format_data, $config);
 
 		$this->assertInstanceOf('\\YoutubeDownloader\\VideoInfo\\Format', $format);
+		$this->assertInstanceOf('\\YoutubeDownloader\\Provider\\Youtube\\Format', $format);
 
 		return $format;
 	}
@@ -63,7 +63,7 @@ class FormatTest extends TestCase
 	 * @test getUrl()
 	 * @depends createFromArray
 	 */
-	public function getUrl(FormatInterface $format)
+	public function getUrl(Format $format)
 	{
 		$this->assertSame(
 			'https://r4---sn-xjpm-q0ns.googlevideo.com/videoplayback?mime=video%2Fmp4&dur=3567.629&id=o-ALLJQEDdJCHuwXcL9toC3Iim4AvgjxD3lzAyP_l8I3i6&pl=23&itag=22&ratebypass=yes&mt=1495797277&ms=au&requiressl=yes&mn=sn-xjpm-q0ns&mm=31&source=youtube&sparams=dur%2Cei%2Cid%2Cinitcwndbps%2Cip%2Cipbits%2Citag%2Clmt%2Cmime%2Cmm%2Cmn%2Cms%2Cmv%2Cpl%2Cratebypass%2Crequiressl%2Csource%2Cexpire&key=yt6&lmt=1471747682170867&signature=B6A1CD71B2245C40BCC26B71BF9DC15545591BE1.DB1653D145BC156D1A3C13F24B63B08BD10F86AB&ei=gQ4oWf3TNsGVceLmsagP&expire=1495818977&ip=211.12.135.54&mv=m&initcwndbps=1286250&ipbits=0',
@@ -75,7 +75,7 @@ class FormatTest extends TestCase
 	 * @test getItag()
 	 * @depends createFromArray
 	 */
-	public function getItag(FormatInterface $format)
+	public function getItag(Format $format)
 	{
 		$this->assertSame('22', $format->getItag());
 	}
@@ -84,7 +84,7 @@ class FormatTest extends TestCase
 	 * @test getQuality()
 	 * @depends createFromArray
 	 */
-	public function getQuality(FormatInterface $format)
+	public function getQuality(Format $format)
 	{
 		$this->assertSame('hd720', $format->getQuality());
 	}
@@ -93,7 +93,7 @@ class FormatTest extends TestCase
 	 * @test getType()
 	 * @depends createFromArray
 	 */
-	public function getType(FormatInterface $format)
+	public function getType(Format $format)
 	{
 		$this->assertSame('video/mp4', $format->getType());
 	}
@@ -102,7 +102,7 @@ class FormatTest extends TestCase
 	 * @test getExpires()
 	 * @depends createFromArray
 	 */
-	public function getExpires(FormatInterface $format)
+	public function getExpires(Format $format)
 	{
 		$this->assertSame('17:16:17 UTC', $format->getExpires());
 	}
@@ -111,7 +111,7 @@ class FormatTest extends TestCase
 	 * @test getIpbits()
 	 * @depends createFromArray
 	 */
-	public function getIpbits(FormatInterface $format)
+	public function getIpbits(Format $format)
 	{
 		$this->assertSame('0', $format->getIpbits());
 	}
@@ -120,7 +120,7 @@ class FormatTest extends TestCase
 	 * @test getIp()
 	 * @depends createFromArray
 	 */
-	public function getIp(FormatInterface $format)
+	public function getIp(Format $format)
 	{
 		$this->assertSame('211.12.135.54', $format->getIp());
 	}
